refactor: split tests into unit and integration directories
