+++ tests/integration/integration.test.js
@@ -5,7 +5,7 @@ import { fileURLToPath } from 'url';
 import { execSync } from 'child_process';
 
 const __dirname = path.dirname(fileURLToPath(import.meta.url));
-const projectRoot = path.join(__dirname, '..');
+const projectRoot = path.join(__dirname, '..', '..');
 const testOutput = path.join(__dirname, '.integration-output');
 
 describe('CLI Integration', () => {
